refactor(editor): extract evaluation into feature (no-changelog)
