+++ packages/frontend/editor-ui/src/components/WorkflowProductionChecklist.vue
@@ -2,7 +2,7 @@
 import { computed, ref, onMounted, watch } from 'vue';
 import { useI18n } from '@n8n/i18n';
 import { useRouter } from 'vue-router';
-import { useEvaluationStore } from '@/stores/evaluation.store.ee';
+import { useEvaluationStore } from '@/features/evaluation.ee/evaluation.store';
 import { useNodeTypesStore } from '@/stores/nodeTypes.store';
 import type { ActionType, WorkflowSettings } from '@/composables/useWorkflowsCache';
 import { useWorkflowSettingsCache } from '@/composables/useWorkflowsCache';
