+++ packages/frontend/editor-ui/src/features/evaluation.ee/components/ListRuns/MetricsChart.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
-import type { TestRunRecord } from '@/api/evaluation.ee';
+import type { TestRunRecord } from '../../evaluation.api';
 import { computed, watchEffect } from 'vue';
 import { Line } from 'vue-chartjs';
-import { useMetricsChart } from '../composables/useMetricsChart';
+import { useMetricsChart } from '../../composables/useMetricsChart';
 
 import { N8nOption, N8nSelect } from '@n8n/design-system';
 const emit = defineEmits<{
