+++ packages/frontend/editor-ui/src/features/evaluation.ee/components/ListRuns/RunsSection.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import type { TestRunRecord } from '@/api/evaluation.ee';
-import MetricsChart from '@/components/Evaluations.ee/ListRuns/MetricsChart.vue';
-import TestRunsTable from '@/components/Evaluations.ee/ListRuns/TestRunsTable.vue';
+import type { TestRunRecord } from '../../evaluation.api';
+import MetricsChart from './MetricsChart.vue';
+import TestRunsTable from './TestRunsTable.vue';
 import { useI18n } from '@n8n/i18n';
 import { VIEWS } from '@/constants';
 import { convertToDisplayDate } from '@/utils/formatters/dateFormatter';
