+++ packages/frontend/editor-ui/src/features/evaluation.ee/components/ListRuns/TestRunsTable.vue
@@ -1,12 +1,11 @@
 <script setup lang="ts">
-import type { TestRunRecord } from '@/api/evaluation.ee';
+import type { TestRunRecord } from '../../evaluation.api';
 import { useI18n } from '@n8n/i18n';
 import { computed } from 'vue';
 import type { TestTableColumn } from '../shared/TestTableBase.vue';
 import type { BaseTextKey } from '@n8n/i18n';
 import TestTableBase from '../shared/TestTableBase.vue';
-import { statusDictionary } from '../shared/statusDictionary';
-import { getErrorBaseKey } from '@/components/Evaluations.ee/shared/errorCodes';
+import { statusDictionary, getErrorBaseKey } from '../../evaluation.constants';
 import { I18nT } from 'vue-i18n';
 import { N8nHeading, N8nIcon, N8nText, N8nTooltip } from '@n8n/design-system';
 import AnimatedSpinner from '@/components/AnimatedSpinner.vue';
