+++ packages/frontend/editor-ui/src/features/evaluation.ee/components/shared/StepHeader.vue
