+++ packages/frontend/editor-ui/src/features/evaluation.ee/components/shared/StepIndicator.vue
