+++ packages/frontend/editor-ui/src/features/evaluation.ee/evaluation.constants.ts
@@ -1,3 +1,7 @@
+import type { TestRunRecord } from './evaluation.api';
+import { type IconName } from '@n8n/design-system/components/N8nIcon/icons';
+import type { IconColor } from '@n8n/design-system/types/icon';
+
 import type { BaseTextKey } from '@n8n/i18n';
 
 const TEST_CASE_EXECUTION_ERROR_CODE = {
@@ -54,4 +58,38 @@ export const getErrorBaseKey = (errorCode?: string): string => {
 		testRunErrorDictionary[errorCode as TestRunErrorCode] ??
 		''
 	);
+};
+
+export const statusDictionary: Record<
+	TestRunRecord['status'],
+	{ icon: IconName; color: IconColor }
+> = {
+	new: {
+		icon: 'status-new',
+		color: 'foreground-xdark',
+	},
+	running: {
+		icon: 'spinner',
+		color: 'secondary',
+	},
+	completed: {
+		icon: 'status-completed',
+		color: 'success',
+	},
+	error: {
+		icon: 'triangle-alert',
+		color: 'danger',
+	},
+	cancelled: {
+		icon: 'status-canceled',
+		color: 'foreground-xdark',
+	},
+	warning: {
+		icon: 'status-warning',
+		color: 'warning',
+	},
+	success: {
+		icon: 'status-completed',
+		color: 'success',
+	},
 };
