+++ packages/frontend/editor-ui/src/features/evaluation.ee/views/EvaluationsRootView.test.ts
@@ -2,16 +2,16 @@ import { describe, it, expect, beforeEach, vi } from 'vitest';
 import { mock } from 'vitest-mock-extended';
 import { createTestingPinia } from '@pinia/testing';
 import { createComponentRenderer } from '@/__tests__/render';
-import EvaluationRootView from '../EvaluationsRootView.vue';
+import EvaluationRootView from './EvaluationsRootView.vue';
 
 import { useWorkflowsStore } from '@/stores/workflows.store';
-import { useEvaluationStore } from '@/stores/evaluation.store.ee';
+import { useEvaluationStore } from '../evaluation.store';
 import { useUsageStore } from '@/stores/usage.store';
 import { useSourceControlStore } from '@/stores/sourceControl.store';
 import { mockedStore } from '@/__tests__/utils';
 import type { IWorkflowDb } from '@/Interface';
 import { waitFor } from '@testing-library/vue';
-import type { TestRunRecord } from '@/api/evaluation.ee';
+import type { TestRunRecord } from '../evaluation.api';
 import { PLACEHOLDER_EMPTY_WORKFLOW_ID } from '@/constants';
 import { useTelemetry } from '@/composables/useTelemetry';
 import { EVALUATION_NODE_TYPE, EVALUATION_TRIGGER_NODE_TYPE, NodeHelpers } from 'n8n-workflow';
